+++ hutool-core/src/main/java/cn/hutool/core/lang/Dict.java
@@ -162,6 +162,7 @@ public class Dict extends HashMap<String, Object> implements BasicTypeGetter<Str
 	 * @return 自己
 	 */
 	public <T> Dict parseBean(T bean) {
+		Assert.notNull(bean, "Bean class must be not null");
 		this.putAll(BeanUtil.beanToMap(bean));
 		return this;
 	}
@@ -177,6 +178,7 @@ public class Dict extends HashMap<String, Object> implements BasicTypeGetter<Str
 	 * @return 自己
 	 */
 	public <T> Dict parseBean(T bean, boolean isToUnderlineCase, boolean ignoreNullValue) {
+		Assert.notNull(bean, "Bean class must be not null");
 		this.putAll(BeanUtil.beanToMap(bean, isToUnderlineCase, ignoreNullValue));
 		return this;
 	}